[GraphQL/Upcast] IOwner.objects returns MoveObjects
Part of the feedback to address from 1.5 -- return a `MoveObject` from
te `objects` connection on owners, because it's impossible for an
object that is owned by another object or an address to be a
package.

At the moment, this makes certain queries more annoying (if you want
to access fields on `Object` from this endpoint), but this
inconvenience will disappear soon, when the explicit upcasts are
replaced by interfaces (in a PR to come).

Test Plan:

Existing E2E and unit tests:

```
sui-graphql-rpc$ cargo nextest run
sui-graphql-e2e-tests$ cargo nextest run \
  -j 1 --features pg_integration
```